requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
